requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.9
django-filter==1.1.0
django-jet==1.0.7
django-rest-swagger==2.1.2
djangorestframework==3.7.7
google-api-python-client==1.4.1
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauth2client==4.1.2
openapi-codec==1.3.2
Pillow==5.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
